Refactor views and security utilities
