+++ root/app/Views/accounts.php
@@ -14,7 +14,7 @@
 
 
 
-require 'layouts/header.php';
+require 'partials/header.php';
 ?>
 
 <main class="container">
@@ -160,4 +160,4 @@ require 'layouts/header.php';
 </script>
 
 <?php
- require 'layouts/footer.php'; ?>
+ require 'partials/footer.php'; ?>
